+++ test/lib/bamboo/postmark_adapter_test.exs
@@ -18,9 +18,9 @@ defmodule Bamboo.PostmarkAdapterTest do
     plug :dispatch
 
     def start_server(parent) do
-      Agent.start_link(fn -> HashDict.new end, name: __MODULE__)
-      Agent.update(__MODULE__, &HashDict.put(&1, :parent, parent))
-      port = get_free_port
+      Agent.start_link(fn -> Map.new end, name: __MODULE__)
+      Agent.update(__MODULE__, &Map.put(&1, :parent, parent))
+      port = get_free_port()
       Application.put_env(:bamboo, :postmark_base_uri, "http://localhost:#{port}")
       Plug.Adapters.Cowboy.http __MODULE__, [], port: port, ref: __MODULE__
     end
@@ -55,14 +55,14 @@ defmodule Bamboo.PostmarkAdapterTest do
     end
 
     defp send_to_parent(conn) do
-      parent = Agent.get(__MODULE__, fn(set) -> HashDict.get(set, :parent) end)
+      parent = Agent.get(__MODULE__, fn(set) -> Map.get(set, :parent) end)
       send parent, {:fake_postmark, conn}
       conn
     end
   end
 
   setup do
-    FakePostmark.start_server(self)
+    FakePostmark.start_server(self())
 
     on_exit fn ->
       FakePostmark.shutdown
@@ -73,7 +73,7 @@ defmodule Bamboo.PostmarkAdapterTest do
 
   test "raises if the api key is nil" do
     assert_raise ArgumentError, ~r/no API key set/, fn ->
-      new_email(from: "foo@bar.com") |> PostmarkAdapter.deliver(@config_with_bad_key)
+      PostmarkAdapter.deliver(new_email(from: "foo@bar.com"), @config_with_bad_key)
     end
 
     assert_raise ArgumentError, ~r/no API key set/, fn ->
@@ -81,8 +81,23 @@ defmodule Bamboo.PostmarkAdapterTest do
     end
   end
 
-  test "deliver/2 sends the to the right url" do
-    new_email |> PostmarkAdapter.deliver(@config)
+  test "deliver/2 passes the request_options to hackney" do
+    request_options = [recv_timeout: 0]
+    config = Map.put(@config, :request_options, request_options)
+
+    assert_raise Bamboo.PostmarkAdapter.ApiError, fn ->
+      PostmarkAdapter.deliver(new_email(), config)
+    end
+  end
+
+  test "deliver/2 returns the textual body of the request" do
+    response = PostmarkAdapter.deliver(new_email(), @config)
+
+    assert response.body == "SENT"
+  end
+
+  test "deliver/2 makes the request to the right url" do
+    PostmarkAdapter.deliver(new_email(), @config)
 
     assert_receive {:fake_postmark, %{request_path: request_path}}
 
@@ -90,7 +105,7 @@ defmodule Bamboo.PostmarkAdapterTest do
   end
 
   test "deliver/2 sends the to the right url for templates" do
-    new_email |> PostmarkHelper.template("hello") |> PostmarkAdapter.deliver(@config)
+    new_email() |> PostmarkHelper.template("hello") |> PostmarkAdapter.deliver(@config)
 
     assert_receive {:fake_postmark, %{request_path: request_path}}
 
@@ -98,23 +113,25 @@ defmodule Bamboo.PostmarkAdapterTest do
   end
 
   test "deliver/2 sends from, html and text body, subject, and headers" do
-    email = new_email(
-      from: {"From", "from@foo.com"},
-      subject: "My Subject",
-      text_body: "TEXT BODY",
-      html_body: "HTML BODY",
-    )
-    |> Email.put_header("Reply-To", "reply@foo.com")
+    email =
+      [
+        from: {"From", "from@foo.com"},
+        subject: "My Subject",
+        text_body: "TEXT BODY",
+        html_body: "HTML BODY",
+      ]
+      |> new_email()
+      |> Email.put_header("Reply-To", "reply@foo.com")
 
-    email |> PostmarkAdapter.deliver(@config)
+    PostmarkAdapter.deliver(email, @config)
 
     assert_receive {:fake_postmark, %{params: params}}
-    assert params["From"] == "#{email.from |> elem(0)} <#{email.from |> elem(1)}>"
+    assert params["From"] == "#{elem(email.from, 0)} <#{elem(email.from, 1)}>"
     assert params["Subject"] == email.subject
     assert params["TextBody"] == email.text_body
     assert params["HtmlBody"] == email.html_body
-    assert params["Headers"] == [%{"Name" => "Reply-To",
-      "Value" => "reply@foo.com"}]
+    assert params["Headers"] ==
+      [%{"Name" => "Reply-To", "Value" => "reply@foo.com"}]
   end
 
   test "deliver/2 correctly formats recipients" do
@@ -124,7 +141,7 @@ defmodule Bamboo.PostmarkAdapterTest do
       bcc: [{"BCC", "bcc@bar.com"}],
     )
 
-    email |> PostmarkAdapter.deliver(@config)
+    PostmarkAdapter.deliver(email, @config)
 
     assert_receive {:fake_postmark, %{params: params}}
     assert params["To"] == "To <to@bar.com>"
@@ -133,22 +150,22 @@ defmodule Bamboo.PostmarkAdapterTest do
   end
 
   test "deliver/2 puts template name and empty content" do
-    email = new_email |> PostmarkHelper.template("hello")
+    email = PostmarkHelper.template(new_email(), "hello")
 
-    email |> PostmarkAdapter.deliver(@config)
+    PostmarkAdapter.deliver(email, @config)
 
     assert_receive {:fake_postmark, %{params: %{"TemplateId" => template_id,
        "TemplateModel" => template_model}}}
     assert template_id == "hello"
-    assert template_model == []
+    assert template_model == %{}
   end
 
   test "deliver/2 puts template name and content" do
-    email = new_email |> PostmarkHelper.template("hello", [
+    email = PostmarkHelper.template(new_email(), "hello", [
       %{name: 'example name', content: 'example content'}
     ])
 
-    email |> PostmarkAdapter.deliver(@config)
+    PostmarkAdapter.deliver(email, @config)
 
     assert_receive {:fake_postmark, %{params: %{"TemplateId" => template_id,
        "TemplateModel" => template_model}}}
@@ -158,11 +175,25 @@ defmodule Bamboo.PostmarkAdapterTest do
   end
 
   test "deliver/2 puts tag param" do
-     email = new_email |> PostmarkHelper.tag("some_tag")
+    email = PostmarkHelper.tag(new_email(), "some_tag")
 
-     email |> PostmarkAdapter.deliver(@config)
+    PostmarkAdapter.deliver(email, @config)
 
-     assert_receive {:fake_postmark, %{params: %{"Tag" => "some_tag"}}}
+    assert_receive {:fake_postmark, %{params: %{"Tag" => "some_tag"}}}
+  end
+
+  test "deliver/2 puts tracking params" do
+    email =
+      new_email()
+      |> PostmarkHelper.template("hello")
+      |> PostmarkHelper.put_param("TrackOpens", true)
+      |> PostmarkHelper.put_param("TrackLinks", "HtmlOnly")
+
+    PostmarkAdapter.deliver(email, @config)
+
+    assert_receive {:fake_postmark, %{params: %{
+      "TrackLinks" => "HtmlOnly", "TrackOpens" => true, "TemplateId" => "hello"}
+    }}
   end
 
   test "deliver/2 puts reply to param" do
@@ -177,12 +208,14 @@ defmodule Bamboo.PostmarkAdapterTest do
     email = new_email(from: "INVALID_EMAIL")
 
     assert_raise Bamboo.PostmarkAdapter.ApiError, fn ->
-      email |> PostmarkAdapter.deliver(@config)
+      PostmarkAdapter.deliver(email, @config)
     end
   end
 
   defp new_email(attrs \\ []) do
-    attrs = Keyword.merge([from: "foo@bar.com", to: []], attrs)
-    Email.new_email(attrs) |> Bamboo.Mailer.normalize_addresses
+    [from: "foo@bar.com", to: []]
+    |> Keyword.merge(attrs)
+    |> Email.new_email()
+    |> Bamboo.Mailer.normalize_addresses()
   end
 end