+++ CHANGELOG.md
@@ -1,3 +1,44 @@
+## 0.4.2
+
+### Fixes/Enhancements
+
+* Relax bamboo and hackney library requirements ([#18])
+* Fix bug when leaving default `template_model` value in `PostmarkHelper.template/3` ([#17])
+* Fix code style issues and compiler warnings ([#16])
+
+[#16]: https://github.com/pablo-co/bamboo_postmark/pull/16
+[#17]: https://github.com/pablo-co/bamboo_postmark/pull/17
+[#18]: https://github.com/pablo-co/bamboo_postmark/pull/18
+
+## 0.4.1
+
+### Fixes/Enhancements
+
+* Return body in response of PostmarkAdapter.deliver ([#14])
+
+[#14]: https://github.com/pablo-co/bamboo_postmark/pull/14
+
+## 0.4.0
+
+### New Additions
+
+* Allow configuration of request options ([#12])
+
+[#12]: https://github.com/pablo-co/bamboo_postmark/pull/12
+
+## 0.3.0
+
+### New Additions
+
+* Support passing custom params to Postmark ([#9])
+
+### Fixes/Enhancements
+
+* Fix Elixir 1.4 warnings and deprecations ([#7])
+
+[#9]: https://github.com/pablo-co/bamboo_postmark/pull/9
+[#7]: https://github.com/pablo-co/bamboo_postmark/pull/7
+
 ## 0.2.0
 
 ### New Additions